♻️ refactor: Print formatted output in printOddOrEven function
This commit refactors the printOddOrEven function in evenorodd.go to improve the formatting of the output. Instead of using the "%02v\t ===> \t %02v" format, it now uses "%v\t===> %s" format to display the values and their respective odd or even status. This change enhances the readability of the printed output.